Enemy instantiation

+++ scripts/battle/entity/character/aj/Aj.kt
@@ -7,10 +7,10 @@ import battle.entity.character.AbstractCharacter
  *
  */
 class Aj : AbstractCharacter<AjNode>(AjNode()) {
-	override var maxHealth: Int = 20
-	override var health: Int = 20
-	override var maxShield: Int = 62
-	override var shield: Int = 62
+	override var maxHealth: Double = 20.0
+	override var health: Double = 20.0
+	override var maxShield: Double = 62.0
+	override var shield: Double = 62.0
 
 	override fun sustainDamage(damage: Double): Double {
 		TODO("Not yet implemented")
