+++ scripts/battle/entity/character/dogman/Dogman.kt
@@ -3,10 +3,10 @@ package battle.entity.character.dogman
 import battle.entity.character.AbstractCharacter
 
 class Dogman : AbstractCharacter<DogmanNode>(DogmanNode()) {
-	override var maxHealth: Int = 36
-	override var health: Int = 36
-	override var maxShield: Int = 15
-	override var shield: Int = 15
+	override var maxHealth: Double = 36.0
+	override var health: Double = 36.0
+	override var maxShield: Double = 15.0
+	override var shield: Double = 15.0
 
 	override fun sustainDamage(damage: Double): Double {
 		TODO("Not yet implemented")
