+++ scripts/battle/entity/character/kew/Kew.kt
@@ -3,10 +3,10 @@ package battle.entity.character.kew
 import battle.entity.character.AbstractCharacter
 
 class Kew : AbstractCharacter<KewNode>(KewNode()) {
-	override var maxHealth: Int = 33
-	override var health: Int = 33
-	override var maxShield: Int = 17
-	override var shield: Int = 17
+	override var maxHealth: Double = 33.0
+	override var health: Double = 33.0
+	override var maxShield: Double = 17.0
+	override var shield: Double = 17.0
 
 
 	override fun sustainDamage(damage: Double): Double {
