+++ scripts/battle/entity/character/peek/Peek.kt
@@ -3,10 +3,10 @@ package battle.entity.character.peek
 import battle.entity.character.AbstractCharacter
 
 class Peek : AbstractCharacter<PeekNode>(PeekNode()) {
-	override var maxHealth: Int = 62
-	override var health: Int = 62
-	override var maxShield: Int = 17
-	override var shield: Int = 17
+	override var maxHealth: Double = 62.0
+	override var health: Double = 62.0
+	override var maxShield: Double = 17.0
+	override var shield: Double = 17.0
 
 	override fun sustainDamage(damage: Double): Double {
 		TODO("Not yet implemented")
